Added Sinatra-Partials
